internship/deep-fake-detection-1/requirements.txt
torch
facenet-pytorch
opencv-python
scikit-learn
numpy
joblib
pillow
torchvision